added check for compareArrays function

+++ src/main/java/gb/jdk/Calculator.java
@@ -8,10 +8,10 @@ public class Calculator {
      * @param second второе число
      * @param <T> тип, ограниченный Number
      * @param <V>> тип, ограниченный Number
-     * @return double
+     * @return float
      */
-    public static<T extends Number, V extends Number> double sum(T first, V second) {
-        return  first.doubleValue() + second.doubleValue();
+    public static<T extends Number, V extends Number> float sum(T first, V second) {
+        return  first.floatValue() + second.floatValue();
     }
 
     /**
@@ -20,10 +20,10 @@ public class Calculator {
      * @param second второе число
      * @param <T> тип, ограниченный Number
      * @param <V>> тип, ограниченный Number
-     * @return double
+     * @return float
      */
-    public static<T extends Number, V extends Number> double multiply(T first, V second) {
-        return first.doubleValue() * second.doubleValue();
+    public static<T extends Number, V extends Number> float multiply(T first, V second) {
+        return first.floatValue() * second.floatValue();
     }
 
     /**
@@ -32,13 +32,13 @@ public class Calculator {
      * @param second второе число
      * @param <T> тип, ограниченный Number
      * @param <V>> тип, ограниченный Number
-     * @return double
+     * @return float
      */
-    public static<T extends Number, V extends Number> double divide(T first, V second) {
-        if (second.doubleValue() == 0) {
+    public static<T extends Number, V extends Number> float divide(T first, V second) {
+        if (second.floatValue() == 0) {
             throw new RuntimeException("Попытка деления на ноль");
         }
-        return first.doubleValue() / second.doubleValue();
+        return first.floatValue() / second.floatValue();
     }
 
     /**
@@ -47,10 +47,10 @@ public class Calculator {
      * @param second второе число
      * @param <T> тип, ограниченный Number
      * @param <V>> тип, ограниченный Number
-     * @return double
+     * @return float
      */
-    public static<T extends Number, V extends Number> double subtract(T first, V second) {
-        return first.doubleValue() - second.doubleValue();
+    public static<T extends Number, V extends Number> float subtract(T first, V second) {
+        return first.floatValue() - second.floatValue();
     }
 
     /**
